+++ src/main/java/it/units/sdm/project/core/FreedomBoard.java
@@ -1,5 +1,7 @@
-package it.units.sdm.project;
+package it.units.sdm.project.core;
 
+import it.units.sdm.project.Position;
+import it.units.sdm.project.Stone;
 import it.units.sdm.project.exceptions.InvalidBoardSizeException;
 import it.units.sdm.project.exceptions.InvalidPositionException;
 import it.units.sdm.project.interfaces.Board;
