+++ src/main/java/it/units/sdm/project/core/FreedomGame.java
@@ -1,5 +1,8 @@
-package it.units.sdm.project;
+package it.units.sdm.project.core;
 
+import it.units.sdm.project.Player;
+import it.units.sdm.project.Position;
+import it.units.sdm.project.TextInput;
 import it.units.sdm.project.interfaces.Game;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
