+++ src/test/java/FreedomLineTests.java
@@ -1,4 +1,4 @@
-import it.units.sdm.project.FreedomLine;
+import it.units.sdm.project.core.FreedomLine;
 import it.units.sdm.project.Position;
 import it.units.sdm.project.Stone;
 import org.junit.jupiter.api.Assertions;
